+++ galsim/sed.py
@@ -640,7 +640,12 @@ class SED(object):
         if self.dimensionless:
             raise TypeError("Cannot calculate flux of dimensionless SED.")
         if bandpass is None: # do bolometric flux
-            return galsim.integ.int1d(self, 0.0, float('inf'))
+            from galsim.deprecated import depr
+            depr('Using calculateFlux(bandpass=None) to compute a bolometric flux', 1.5, '',
+                 "If you need this functionality, you can use a pseudo-bolometric Bandpass created "
+                 "with:  bp = Bandpass('1', blue_limit=sed.blue_limit, red_limit=sed.red_limit)")
+            bp = galsim.Bandpass('1', self.blue_limit, self.red_limit)
+            return self.calculateFlux(bp)
         else: # do flux through bandpass
             if len(bandpass.wave_list) > 0 or len(self.wave_list) > 0:
                 slop = 1e-6 # nm